read pkg energy

+++ rapl-lib/src/rapl.rs
@@ -54,9 +54,9 @@ pub mod amd {
     the "Cores" energy usage separately for each core, rather than a
     per-package total"
      */
-    pub const MSR_RAPL_POWER_UNIT: u32 = 0xC0010299; // Similar to Intel MSR_RAPL_POWER_UNIT
-    pub const AMD_MSR_CORE_ENERGY: u32 = 0xC001029A; // Similar to Intel PP0_ENERGY_STATUS (PP1 is for the GPU)
-    pub const MSR_RAPL_PKG_ENERGY_STAT: u32 = 0xC001029B; // Similar to Intel PKG_ENERGY_STATUS (This is for the whole socket)
+    pub const MSR_RAPL_POWER_UNIT: u64 = 0xC0010299; // Similar to Intel MSR_RAPL_POWER_UNIT
+    pub const AMD_MSR_CORE_ENERGY: u64 = 0xC001029A; // Similar to Intel PP0_ENERGY_STATUS (PP1 is for the GPU)
+    pub const MSR_RAPL_PKG_ENERGY_STAT: u64 = 0xC001029B; // Similar to Intel PKG_ENERGY_STATUS (This is for the whole socket)
 
     /*
     const AMD_TIME_UNIT_MASK: u64 = 0xF0000;
